data description of spam email
we handle a data set relative to some measures on n = 4601 mails labeled as "mail" or "spam". This data set was first considered for Hewlett-Packard technical report and is now famous as a training/testing set. The aim is to determine if a mail is spam or not. To do this, the data set contains p = 57 real variables x1, . . . , x57 and one qualitative variable (0 for mail and 1 for spam). The data set contains 1813 spams (i.e. 39.4% of spams). The variables measures various things as the frequencies of some particular words ("make", "address", "george" , . . . ), the frequencies of some characters (’;’, ’$’, ’#’, . . . ), lengths of uninterrupted sequences of capital letters, . . .