data description of  women contraceptive methods
This dataset is a subset of the 1987 National Indonesia Contraceptive Prevalence Survey. It is posted on UCI data mining websites. The samples are married women. The problem is to predict the current contraceptive method choice (no use, long-term methods, or short-term methods) of a woman based on her demographic and socio-economic characteristics.
It has 9 predict attributes and 1 response attribute. most of the predict attributes are defined by categories(e.g. from 1 to 4, 1 is low, 4 is high), and it is a classification case. No missing values.